+++ packages/atomic/cypress/integration-insight-panel/insight-panel.cypress.ts
@@ -37,24 +37,6 @@ describe('Insight Panel test suites', () => {
         .should('have.length.at.least', 1);
     });
 
-    it('should not add any unexpected style tags', () => {
-      const numTopLevelStyleTags = 4;
-      const numLayoutStyleTags = 1;
-
-      cy.get('style').should(
-        'have.length',
-        numTopLevelStyleTags + numLayoutStyleTags
-      );
-      InsightPanelsSelectors.topLevelStyleTags().should(
-        'have.length',
-        numTopLevelStyleTags
-      );
-      InsightPanelsSelectors.layoutStyleTags().should(
-        'have.length',
-        numLayoutStyleTags
-      );
-    });
-
     it('should display results', () => {
       InsightPanelsSelectors.results()
         .should('exist')
@@ -92,13 +74,6 @@ describe('Insight Panel test suites', () => {
         .should('have.attr', 'value', '2');
     });
 
-    it('should display query summary', () => {
-      InsightPanelsSelectors.querySummary()
-        .should('exist')
-        .shadow()
-        .should('have.text', 'Insights related to this case');
-    });
-
     it('should display a search box', () => {
       InsightPanelsSelectors.searchbox()
         .should('exist')
@@ -122,36 +97,6 @@ describe('Insight Panel test suites', () => {
         .should('have.attr', 'title', 'Filters');
     });
 
-    it('should display refine-modal', () => {
-      InsightPanelsSelectors.refineToggle().click();
-      InsightPanelsSelectors.refineModal().should('exist');
-      InsightPanelsSelectors.focusTrap().should('exist');
-
-      InsightPanelsSelectors.filtersModal().should('have.attr', 'is-open');
-
-      InsightPanelsSelectors.filtersModal()
-        .find('[slot~="header"]')
-        .should('have.text', 'Filters');
-
-      InsightPanelsSelectors.filters()
-        .find('atomic-insight-facet')
-        .should('be.visible')
-        .should('have.length.at.least', 1);
-
-      InsightPanelsSelectors.filters()
-        .find('atomic-insight-timeframe-facet')
-        .should('be.visible')
-        .should('have.length.at.least', 1);
-
-      InsightPanelsSelectors.filters()
-        .find('atomic-insight-numeric-facet')
-        .should('be.visible')
-        .should('have.length.at.least', 1);
-
-      InsightPanelsSelectors.filtersModal().find('[slot="footer"]').click();
-
-      InsightPanelsSelectors.filtersModal().should('not.have.attr', 'is-open');
-    });
 
     it('should display edit toggle', () => {
       InsightPanelsSelectors.editToggle()
@@ -225,7 +170,6 @@ describe('Insight Panel test suites', () => {
     });
   });
 
-
   describe('Smart Snippet Answer', () => {
     const visitPage = () => {
       cy.visit(host);
@@ -262,19 +206,6 @@ describe('Insight Panel test suites', () => {
 
       it('should show the smart snippet component', () => {
         InsightPanelsSelectors.smartSnippetExpandableAnswer().should('exist');
-      });
-
-      describe('when giving explanatory feedback', () => {
-        it('should show the feedback modal', () => {
-          InsightPanelsSelectors.smartSnippetFeedbackModal().should(
-            'not.exist'
-          );
-
-          InsightPanelsSelectors.smartSnippetFeedbackNoButton().click();
-          InsightPanelsSelectors.smartSnippetsExplainWhyButton().click();
-
-          InsightPanelsSelectors.smartSnippetFeedbackModal().should('exist');
-        });
       });
     });
   });