+++ packages/atomic/src/components/common/lazy-index.ts
@@ -6,6 +6,8 @@ export default {
     await import('./atomic-component-error/atomic-component-error.js'),
   'atomic-facet-date-input': async () =>
     await import('./atomic-facet-date-input/atomic-facet-date-input.js'),
+  'atomic-facet-number-input': async () =>
+    await import('./atomic-facet-number-input/atomic-facet-number-input.js'),
   'atomic-facet-placeholder': async () =>
     await import('./atomic-facet-placeholder/atomic-facet-placeholder.js'),
   'atomic-focus-trap': async () =>
@@ -16,6 +18,28 @@ export default {
   'atomic-modal': async () => await import('./atomic-modal/atomic-modal.js'),
   'atomic-numeric-range': async () =>
     await import('./atomic-numeric-range/atomic-numeric-range.js'),
+  'atomic-result-placeholder': async () =>
+    await import('./atomic-result-placeholder/atomic-result-placeholder.js'),
+  'atomic-smart-snippet-answer': async () =>
+    await import(
+      './atomic-smart-snippet-answer/atomic-smart-snippet-answer.js'
+    ),
+  'atomic-smart-snippet-collapse-wrapper': async () =>
+    await import(
+      './atomic-smart-snippet-collapse-wrapper/atomic-smart-snippet-collapse-wrapper.js'
+    ),
+  'atomic-smart-snippet-expandable-answer': async () =>
+    await import(
+      './atomic-smart-snippet-expandable-answer/atomic-smart-snippet-expandable-answer.js'
+    ),
+  'atomic-tab-bar': async () =>
+    await import('./atomic-tab-bar/atomic-tab-bar.js'),
+  'atomic-tab-button': async () =>
+    await import('./atomic-tab-button/atomic-tab-button.js'),
+  'atomic-tab-popover': async () =>
+    await import('./atomic-tab-popover/atomic-tab-popover.js'),
+  'atomic-timeframe': async () =>
+    await import('./atomic-timeframe/atomic-timeframe.js'),
 } as Record<string, () => Promise<unknown>>;
 
 export type * from './index.js';